refactor: 채널 API와 서비스의 DTO 참조 경로를 정리
