+++ src/main/java/com/example/inflace/domain/channel/controller/ChannelApi.java
@@ -1,17 +1,14 @@
 package com.example.inflace.domain.channel.controller;
 
-import com.example.inflace.domain.channel.dto.ChannelEngagementRateResponse;
-import com.example.inflace.domain.channel.dto.ChannelKpiResponse;
-import com.example.inflace.domain.channel.dto.ChannelNewSubscriberResponse;
-
-import com.example.inflace.domain.channel.dto.ChannelTopMainVideosResponse;
-import com.example.inflace.domain.channel.dto.ChannelSubscriberDistributionResponse;
-import com.example.inflace.domain.channel.dto.ChannelSubscriberPatternResponse;
-
-import com.example.inflace.domain.channel.dto.ChannelSubscriberTrendResponse;
-import com.example.inflace.domain.channel.dto.ChannelTopVideosResponse;
-import com.example.inflace.domain.channel.dto.ChannelVideosResponse;
-
+import com.example.inflace.domain.channel.dto.response.ChannelEngagementRateResponse;
+import com.example.inflace.domain.channel.dto.response.ChannelKpiResponse;
+import com.example.inflace.domain.channel.dto.response.ChannelNewSubscriberResponse;
+import com.example.inflace.domain.channel.dto.response.ChannelSubscriberDistributionResponse;
+import com.example.inflace.domain.channel.dto.response.ChannelSubscriberPatternResponse;
+import com.example.inflace.domain.channel.dto.response.ChannelSubscriberTrendResponse;
+import com.example.inflace.domain.channel.dto.response.ChannelTopMainVideosResponse;
+import com.example.inflace.domain.channel.dto.response.ChannelTopVideosResponse;
+import com.example.inflace.domain.channel.dto.response.ChannelVideosResponse;
 import com.example.inflace.global.exception.ApiErrorDefines;
 import com.example.inflace.global.exception.ErrorDefine;
 import com.example.inflace.global.response.BaseResponse;
