+++ src/main/java/com/example/inflace/domain/channel/service/ChannelService.java
@@ -3,22 +3,23 @@ package com.example.inflace.domain.channel.service;
 import com.example.inflace.domain.channel.domain.Channel;
 import com.example.inflace.domain.channel.domain.ChannelAnalytics;
 import com.example.inflace.domain.channel.domain.ChannelStats;
-import com.example.inflace.domain.channel.dto.ChannelTopMainVideosResponse;
 import com.example.inflace.domain.channel.domain.SubscriberLog;
-import com.example.inflace.domain.channel.dto.ChannelEngagementRateResponse;
-import com.example.inflace.domain.channel.dto.ChannelKpiResponse;
-import com.example.inflace.domain.channel.dto.ChannelNewSubscriberResponse;
-import com.example.inflace.domain.channel.dto.ChannelNewSubscriberResponse.NewSubscriberVideo;
-import com.example.inflace.domain.channel.dto.ChannelSubscriberDistributionResponse;
-import com.example.inflace.domain.channel.dto.ChannelSubscriberPatternResponse;
-import com.example.inflace.domain.channel.dto.ChannelSubscriberTrendResponse;
 import com.example.inflace.domain.channel.dto.ChannelVideoSliceResult;
-import com.example.inflace.domain.channel.dto.enums.ChannelSubscriberTrendRange;
-import com.example.inflace.domain.channel.dto.enums.ChannelVideoFormat;
-import com.example.inflace.domain.channel.dto.enums.ChannelVideoSort;
-import com.example.inflace.domain.channel.dto.ChannelVideosRequest;
-import com.example.inflace.domain.channel.dto.ChannelVideosResponse;
-import com.example.inflace.domain.channel.dto.YoutubeDataChannelResponse;
+import com.example.inflace.domain.channel.dto.request.ChannelSubscriberTrendRange;
+import com.example.inflace.domain.channel.dto.request.ChannelVideoFormat;
+import com.example.inflace.domain.channel.dto.request.ChannelVideoSort;
+import com.example.inflace.domain.channel.dto.request.ChannelVideosRequest;
+import com.example.inflace.domain.channel.dto.response.ChannelEngagementRateResponse;
+import com.example.inflace.domain.channel.dto.response.ChannelKpiResponse;
+import com.example.inflace.domain.channel.dto.response.ChannelNewSubscriberResponse;
+import com.example.inflace.domain.channel.dto.response.ChannelNewSubscriberResponse.NewSubscriberVideo;
+import com.example.inflace.domain.channel.dto.response.ChannelSubscriberDistributionResponse;
+import com.example.inflace.domain.channel.dto.response.ChannelSubscriberPatternResponse;
+import com.example.inflace.domain.channel.dto.response.ChannelSubscriberTrendResponse;
+import com.example.inflace.domain.channel.dto.response.ChannelTopMainVideosResponse;
+import com.example.inflace.domain.channel.dto.response.ChannelTopVideosResponse;
+import com.example.inflace.domain.channel.dto.response.ChannelVideosResponse;
+import com.example.inflace.domain.channel.dto.response.YoutubeDataChannelResponse;
 import com.example.inflace.domain.channel.repository.ChannelRepository;
 import com.example.inflace.domain.channel.repository.ChannelAnalyticsRepository;
 import com.example.inflace.domain.channel.repository.SubscriberLogRepository;
@@ -26,7 +27,6 @@ import com.example.inflace.domain.channel.repository.ChannelStatsRepository;
 import com.example.inflace.domain.video.domain.Video;
 import com.example.inflace.domain.video.domain.VideoAnalytics;
 import com.example.inflace.domain.video.domain.VideoStats;
-import com.example.inflace.domain.channel.dto.ChannelTopVideosResponse;
 import com.example.inflace.domain.video.dto.VideoType;
 import com.example.inflace.domain.video.repository.VideoQueryRepository;
 import com.example.inflace.domain.video.repository.VideoAnalyticsRepository;
