+++ src/main/java/com/example/inflace/domain/video/repository/VideoQueryRepository.java
@@ -1,7 +1,7 @@
 package com.example.inflace.domain.video.repository;
 
 import com.example.inflace.domain.channel.dto.ChannelVideoSliceResult;
-import com.example.inflace.domain.channel.dto.ChannelVideosRequest;
+import com.example.inflace.domain.channel.dto.request.ChannelVideosRequest;
 
 public interface VideoQueryRepository {
     ChannelVideoSliceResult findChannelVideos(Long channelId, ChannelVideosRequest request);
