+++ src/main/java/com/example/inflace/global/client/YoutubeDataApiClient.java
@@ -1,6 +1,6 @@
 package com.example.inflace.global.client;
 
-import com.example.inflace.domain.channel.dto.YoutubeDataChannelResponse;
+import com.example.inflace.domain.channel.dto.response.YoutubeDataChannelResponse;
 import com.example.inflace.domain.video.dto.YoutubeDataVideoResponse;
 import com.example.inflace.global.properties.YoutubeProperties;
 import java.net.URI;
